# packages/client/openapi/stim-server.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "stim-server",
    "description": "",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-dev.0"
  },
  "paths": {
...
    "/api/v1/agents/instances/{instance_id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "get_agent_instance",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Agent runtime instance identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registered agent instance projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentInstanceRecord"
                }
              }
            }
          },
          "404": {
            "description": "Agent instance not registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "agents"
        ],
        "operationId": "register_agent_instance",
        "parameters": [
          {
            "name": "instance_id",
            "in": "path",
            "description": "Agent runtime instance identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInstanceRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Registered agent instance projection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentInstanceRecord"
                }
              }
            }
          },
          "400": {
            "description": "Path/body mismatch",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentInstanceRecord": {
        "type": "object",
        "required": [
          "agent_id",
          "instance_id",
          "participant_id",
          "label",
          "agent_kind",
          "capabilities",
          "status",
          "registered_at",
          "last_seen_at",
          "last_event_id"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "agent_kind": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "delivery_endpoint_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance_id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "last_event_id": {
            "type": "string"
          },
          "last_seen_at": {
            "type": "string"
          },
          "participant_id": {
            "type": "string"
          },
          "profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "registered_at": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AgentInstanceStatus"
          }
        }
      },
      "AgentInstanceRegistrationRequest": {
        "type": "object",
        "required": [
          "agent_id",
          "instance_id",
          "label",
          "agent_kind",
          "capabilities",
          "status"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "agent_kind": {
            "type": "string"
          },
          "capabilities": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "delivery_endpoint_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoint": {
            "type": [
              "string",
              "null"
            ]
          },
          "instance_id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "participant_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/AgentInstanceStatus"
          }
        }
      },
      "AgentInstanceStatus": {
        "type": "string",
        "enum": [
          "ready",
          "degraded",
          "unreachable",
          "offline"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}